fixed 64bit compile warnings

+++ AgilityBook/src/Win/AgilityBookCalendarListView.cpp
@@ -1063,7 +1063,7 @@ bool CAgilityBookCalendarListView::OnCmd(int id)
 					std::vector<wxString> line;
 					GetPrintLine(-1, line);
 					table.StartLine();
-					for (size_t i = 0; i < line.size(); ++i)
+					for (int i = 0; i < static_cast<int>(line.size()); ++i)
 					{
 						table.Cell(i, line[i]);
 					}
@@ -1087,7 +1087,7 @@ bool CAgilityBookCalendarListView::OnCmd(int id)
 					std::vector<wxString> line;
 					GetPrintLine((*iter), line);
 					table.StartLine();
-					for (size_t i = 0; i < line.size(); ++i)
+					for (int i = 0; i < static_cast<int>(line.size()); ++i)
 					{
 						table.Cell(i, line[i]);
 					}
